--- examples/resources/materialize_source/import.sh
@@ -1,2 +1,0 @@
-# Sources can be imported using the source id:
-terraform import materialize_source.example_source_load_generator <source_id>
