+++ examples/resources/materialize_source_kafka/import.sh
@@ -1,0 +1,2 @@
+# Sources can be imported using the source id:
+terraform import materialize_source_kafka.example_source_kafka <source_id>
