+++ examples/resources/materialize_source_loadgen/import.sh
@@ -1,0 +1,2 @@
+# Sources can be imported using the source id:
+terraform import materialize_source_loadgen.example_source_load_generator <source_id>
