+++ examples/resources/materialize_source_postgres/import.sh
@@ -1,0 +1,2 @@
+# Sources can be imported using the source id:
+terraform import materialize_source_postgres.example_source_postgres <source_id>
